Added Hero bar. Updated Navigation. Added hover and other animation. Updated policies.
Added hero bar for top of page with falling leaves animation and changing gradient bg. Going to add
parallax for when images are placed on it.
Updated Nav to have "Buy Tickets Now" button and date at the top.
Updated harassment and privacy policies to read from markdown and display correctly.